backend/requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
geopy>=2.0.0
requests>=2.0.0
pyswisseph>=2.10.3
